Move Card components to onboarding
