+++ apps/onboarding/src/components/Card/Card.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
-import { Space } from '../Space'
+import { Space } from 'ui/src/components/Space'
 import { Card, CardContent, CardMedia } from './Card'
 
 export default {
